Rebase/requirements.txt
firebase_admin==6.3.0
python-socketio==5.10.0
python-dotenv==1.0.0
redis==5.0.1
aioredis==2.0.1
fastapi==0.108.0
pydantic==2.5.3
uvicorn==0.25.0
requests==2.31.0
python-docx==1.1.0
google-generativeai==0.3.2
